util/log: move the canonical format functions to a new file
Release note: None
